% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP.Subcortical.R
\docType{data}
\name{EP.Subcortical}
\alias{EP.Subcortical}
\title{Expected patterns of subcortical}
\format{
A data frame with 8 rows and 4 variables:
\describe{
\item{Subcortical}{names of regional neuroimaging traits}
\item{SSD}{the expected pattern of Schizophrenia Spectrum Disorder}
\item{MDD}{the expected pattern of Major Depressive Disorder}
\item{AD}{the expected pattern of Alzheimer's Disease}
}
}
\source{
The 'EP.Subcortical' pattern is from Maryland Psychiatric Research Center, Department of Psychiatry, University of Maryland School of Medicine.
}
\usage{
EP.Subcortical
}
\description{
the expected patterns of subcortical in schizophrenia derived from large-scale meta-analyses by the ENIGMA.
}
\keyword{datasets}
