% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{test}
\alias{test}
\title{test}
\format{
A data frame with 196 rows and 12 variables:
\describe{
\item{ID}{subjects' ID}
\item{DX}{indicators of control group}
\item{Age}{subjects' age}
\item{Sex}{subjects' gehder}
\item{Lateral.Ventricle}{simulated volumes of lateral Ventricle}
\item{Thalamus}{simulated volumes of Thalamus}
\item{Caudate}{simulated volumes of Caudate}
\item{Putamen}{simulated volumes of Putamen}
\item{Pallidum}{simulated volumes of Pallidum}
\item{Hippocampus}{simulated volumes of Hippocampus}
\item{Amygdala}{simulated volumes of Amygdala}
\item{Left.Accumbens.area}{simulated volumes of left Accumbens}
}
}
\source{
The 'test' dataset is from Maryland Psychiatric Research Center, Department of Psychiatry, University of Maryland School of Medicine.
}
\usage{
test
}
\description{
Simulated volumes of subcortical structures of Schizophrenia spectrum disorder are used as an example for this function.
You can calculate RVI for this dataset"
}
\keyword{datasets}
