% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP.GM.R
\docType{data}
\name{EP.GM}
\alias{EP.GM}
\title{Expected patterns of gray matter}
\format{
A data frame with 33 rows and 11 variables:
\describe{
\item{GM}{grey matter thickness}
\item{SSD}{the expected pattern of Schizophrenia Spectrum Disorder (\doi{10.1016/j.biopsych.2018.04.023} table S4a)}
\item{MDD}{the expected pattern of Major Depressive Disorder (\doi{10.1038/mp.2016.60} table 1)}
\item{AD_ADNI}{the expected pattern of Alzheimer's Disease (\doi{10.1016/j.nicl.2021.102574} table S1)}
\item{AD_ADNIOSYRIX}{the expected pattern of Alzheimer's Disease Neuroimaging Initiative + OSIRIX}
\item{BD}{the expected pattern of Bipolar Disorder (\doi{10.1038/mp.2017.73} table 1)}
\item{PD}{the expected pattern of Parkinson's disease}
\item{Diabetes}{the expected pattern of Diabetes}
\item{HighBP}{the expected pattern of Hypertension}
\item{HighLipids}{the expected pattern of Hyperlipidemia}
\item{MET}{the expected pattern of Metabolic diseases(Diabetes, Hypertension and Hyperlipidemia)}
}
}
\source{
The 'EP.GM' pattern is from Maryland Psychiatric Research Center, Department of Psychiatry, University of Maryland School of Medicine.
}
\usage{
EP.GM
}
\description{
the expected patterns of gray matter in schizophrenia derived from large-scale meta-analyses by the ENIGMA.
}
\keyword{datasets}
