% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP.WM.R
\docType{data}
\name{EP.WM}
\alias{EP.WM}
\title{Expected patterns of white matter}
\format{
A data frame with 25 rows and 5 variables:
\describe{
\item{WM}{white matter fractional anisotropy}
\item{SSD}{the expected pattern of Schizophrenia Spectrum Disorder}
\item{MDD}{the expected pattern of Major Depressive Disorder}
\item{AD}{the expected pattern of Alzheimer's Disease}
\item{BD}{the expected pattern of Bipolar Disorder}
\item{ADNI_OSIRIX}{the expected pattern of Alzheimer's Disease Neuroimaging Initiative + OSIRIX}
\item{Diabetes}{the expected pattern of diabetes}
\item{HighBP}{the expected pattern of high blood pressure (hypertension)}
\item{HighLipids}{the expected pattern of hyperlipidemia}
}
}
\source{
The 'EP.WM' pattern is from Maryland Psychiatric Research Center, Department of Psychiatry, University of Maryland School of Medicine.
}
\usage{
EP.WM
}
\description{
the expected patterns of white matter in schizophrenia derived from large-scale meta-analyses by the ENIGMA.
}
\keyword{datasets}
